+++ src/utils/config.ts
@@ -24,9 +24,5 @@ export const localConfig = {
 			name: './projects',
 			alias: 'projects',
 		},
-		{
-			name: './run',
-			alias: 'run',
-		},
 	],
 };
